requirements.txt
fastapi
asgi_correlation_id
httpx
aiofile
openpyxl
python-docx
faiss-cpu
jieba
langchain
langchain-community
openai
pillow
opencv-python
onnxruntime
pyclipper
shapely
pymupdf
gradio
requests==2.32.3
pypdf
beautifulsoup4
